Add check whether backup is supported
If source and target database are incompatible SQLite's backup silently fails and leaves a corrupted database behind.

Databases are compatible with SQLite's backup, if they are both plain unencrypted databases or if they are both encrypted with ciphers that have at least the same number of reserved bytes per page.